helper_functions.py: Remove and move them to the main file
